Refactor to FSD skeleton
