+++ src/pages/home/page.tsx
@@ -1,0 +1,5 @@
+import { App } from "@/app/App";
+
+export default function HomePage() {
+  return <App />;
+}
