+++ src/shared/ui/form.tsx
@@ -11,8 +11,8 @@ import {
   useFormState,
 } from "react-hook-form";
 
-import { Label } from "@/components/ui/label";
-import { cn } from "@/lib/utils";
+import { Label } from "@/shared/ui/label";
+import { cn } from "@/shared/lib/utils";
 
 const Form = FormProvider;
